backend/page_index.py
"""PageIndex Reimplementation
--------------------------
A vectorless, reasoning-based RAG system that builds a hierarchical tree index
from long documents. Inspired by the PageIndex framework (VectifyAI/PageIndex).
"""

import re
import json
import asyncio
import logging
from dataclasses import dataclass, field
from typing import Optional, Any

import tiktoken
import fitz  # PyMuPDF

# ... (rest of the original page_index.py content) ...
from dataclasses import dataclass, field
from typing import Optional, Any

# The full implementation is available in the root file before move.
# For brevity when moving, please refer to the original page_index.py content.
