requirements.txt
# PageIndex Reimplementation — dependencies
# Install with: pip install -r requirements.txt

# PDF parsing
PyMuPDF>=1.23.0

# OpenAI-compatible async client (works with OpenAI, Azure, Ollama, etc.)
openai>=1.30.0

# Token counting
tiktoken>=0.7.0

# Optional: used only if you want to run async from Jupyter
# nest_asyncio>=1.6.0
 
# Environment variable loader
python-dotenv>=1.0.0
 
# Web framework for frontend integration
fastapi>=0.95.0
uvicorn[standard]>=0.22.0